Requests de Matriculas e Cadastros Feitos

+++ src/main/java/basePackage/entities/Modalidade.java
@@ -22,5 +22,5 @@ public class Modalidade {
 
     private String nome;
 
-    private Long limiteAlunos;
+    private int limiteAlunos;
 }
